fix: rename data.json to overlay_data.json, fix mod path lookup
- Game was parsing data.json as a mod manifest (scans all .json files)
- Renamed to overlay_data.json to avoid conflict
- Fixed mod path: use Assembly.Location with AllMods fallback
- Fixed Glicko2RatingResult GamesPlayed int→long for Dapper compat
- Updated ExportMod to use Glicko2 instead of removed Elo classes

+++ src/Sts2Analytics.Core/Models/AnalyticsResults.cs
@@ -20,7 +20,7 @@ public record CardImpactScore(string CardId, double PickRate, double WinRateDelt
 
 public record Glicko2RatingResult(
     string CardId, string Character, string Context,
-    double Rating, double RatingDeviation, double Volatility, int GamesPlayed);
+    double Rating, double RatingDeviation, double Volatility, long GamesPlayed);
 
 public record Glicko2HistoryResult(
     double RatingBefore, double RatingAfter,
